+++ izneo_get/plugins/site_processor.py
@@ -9,7 +9,13 @@ from tqdm.asyncio import tqdm
 
 from ..book_infos import BookInfos
 from ..config import Config, ImageFormat, OutputFormat
-from ..tools import BAR_FORMAT, async_http_get, clean_name, get_image_type, get_name_from_pattern
+from ..tools import (
+    BAR_FORMAT,
+    async_http_get,
+    clean_name,
+    get_image_type,
+    get_name_from_pattern,
+)
 
 
 class SiteProcessor:
@@ -28,7 +34,9 @@ class SiteProcessor:
 
     @staticmethod
     def is_valid_url(url: str) -> bool:
-        return any(re.match(pattern, url) is not None for pattern in SiteProcessor.URL_PATTERNS)
+        return any(
+            re.match(pattern, url) is not None for pattern in SiteProcessor.URL_PATTERNS
+        )
 
     def authenticate(self) -> None: ...
 
@@ -52,7 +60,9 @@ class SiteProcessor:
 
         if forced_title:
             forced_title = get_name_from_pattern(forced_title, book_infos)
-            print(f'Download "{clean_name(title_used)}" as "{clean_name(forced_title)}"')
+            print(
+                f'Download "{clean_name(title_used)}" as "{clean_name(forced_title)}"'
+            )
             title_used = forced_title
         else:
             print(f'Download "{clean_name(title_used)}"')
@@ -76,7 +86,9 @@ class SiteProcessor:
         if self.config.pause_sec:
             files_downloaded = self._download_all_pages(title_used, save_path)
         else:
-            files_downloaded = asyncio.run(self._async_download_all_pages(title_used, save_path))
+            files_downloaded = asyncio.run(
+                self._async_download_all_pages(title_used, save_path)
+            )
         count_empty = len([element for element in files_downloaded if not element])
         print(f"{len(files_downloaded) - count_empty} pages downloaded")
         if count_empty:
@@ -88,11 +100,18 @@ class SiteProcessor:
 
     def after_download(self, files_downloaded: List[str]) -> None: ...
 
-    def post_process_image_content(self, content: bytes, page_num: int = 0) -> bytes:
-        return content
+    def post_process_image_content(
+        self, response: requests.models.Response, page_num: int = 0
+    ) -> bytes:
+        return response.content
 
     async def _async_download_page(
-        self, page_num: int, url: str, title_used: str, save_path: str, pause_sec: int = 0
+        self,
+        page_num: int,
+        url: str,
+        title_used: str,
+        save_path: str,
+        pause_sec: int = 0,
     ) -> str:
         book_infos = self.get_book_infos()
         if len(book_infos.page_urls) == 0:
@@ -126,14 +145,16 @@ class SiteProcessor:
 
         if r.status_code in (403, 404):
             if page_num < book_infos.pages:
-                print(f"\n[ERROR] Can't download page {str(page_num + 1)} ({str(book_infos.pages)} pages expected)")
+                print(
+                    f"\n[ERROR] Can't download page {str(page_num + 1)} ({str(book_infos.pages)} pages expected)"
+                )
             return ""
         if r.encoding:
             print(f"\n[ERROR] Page {page_num} unavailable")
             return ""
 
         # Decode image.
-        uncrypted = self.post_process_image_content(r.content, page_num=page_num)
+        uncrypted = self.post_process_image_content(r, page_num=page_num)
         store_path = f"{save_path}/{title_used} {page_txt}.tmp"
         open(store_path, "wb").write(uncrypted)
 
@@ -198,7 +219,9 @@ class SiteProcessor:
             os.mkdir(save_path)
         print(f"Destination : {save_path}")
 
-    def create_output_folder(self, book_infos: BookInfos, output_folder: Optional[str]) -> str:
+    def create_output_folder(
+        self, book_infos: BookInfos, output_folder: Optional[str]
+    ) -> str:
         if not output_folder:
             output_folder = "DOWNLOADS"
         output_folder = get_name_from_pattern(output_folder, book_infos)
@@ -207,9 +230,9 @@ class SiteProcessor:
         return output_folder
 
     def _get_title_to_use(self, book_infos: BookInfos) -> str:
-        return get_name_from_pattern(self.config.output_filename or "", book_infos) or self.get_default_title(
-            book_infos
-        )
+        return get_name_from_pattern(
+            self.config.output_filename or "", book_infos
+        ) or self.get_default_title(book_infos)
 
     def get_default_title(self, book_infos: BookInfos) -> str:
         title_used = book_infos.title
@@ -220,7 +243,9 @@ class SiteProcessor:
             if len(volume) > 0:
                 title_used = f'{book_infos.title} - {f"00000{volume}"[-max(2, len(volume)):]}. {subtitle}'
         if len(subtitle) == 0 and len(volume) > 0:
-            title_used = f'{book_infos.title} - {f"00000{volume}"[-max(2, len(volume)):]}'
+            title_used = (
+                f'{book_infos.title} - {f"00000{volume}"[-max(2, len(volume)):]}'
+            )
         return title_used